[TEST] Add unittest util to get available port
Add unittest util to get available port
